requirements.txt
# Web框架和API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# PPT处理
python-pptx==0.6.23

# 图像处理
Pillow==10.2.0

# 语音合成
edge-tts==6.1.10

# 视频处理
moviepy==1.0.3
opencv-python>=4.8.0

# 字幕处理
pysrt==1.1.2
webvtt-py==0.4.6

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# HTTP客户端
requests>=2.31.0
aiohttp>=3.9.0

# 文件处理
aiofiles>=23.2.0

# 日志和配置
python-json-logger>=2.0.7

# 开发工具
python-multipart>=0.0.6
